+++ app/src/app/regionen/[regionSlug]/_components/SidebarInspector/TagsTable/translations/ConditionalFormattedValue.tsx
@@ -171,6 +171,7 @@ export const ConditionalFormattedValue = ({ sourceId, tagKey, tagValue }: Props)
     // 'fresh', // true key is `maxspeed_fresh`, `surface_fresh`, … but we overwrite that when passing props
     'width_source',
     'length',
+    'livecycle',
   ]
   if (!translations[translationKey] && nonCategorizedTagKeys.includes(tagKey)) {
     translationKey = `ALL--${tagKey}=${tagValue}`